agent/requirements/prod.txt
Flask
Flask-RESTful
kafka-python
picamera
RPi.GPIO